+++ src/EliteChroma/Controls/MenuButton.cs
@@ -7,6 +7,8 @@ namespace EliteChroma.Controls
     [ExcludeFromCodeCoverage]
     internal class MenuButton : Button
     {
+        private const char _whiteSpace = '\u2003';
+
         public enum HorizontalDirection
         {
             Right = 0,
@@ -21,6 +23,13 @@ namespace EliteChroma.Controls
 
         [DefaultValue(HorizontalDirection.Right)]
         public HorizontalDirection MenuHorizontalDirection { get; set; }
+
+        // We just insert an em space to the end so the rendered text gets shifted to the left, and avoid rewriting the full render code.
+        public new string Text
+        {
+            get => base.Text.EndsWith(_whiteSpace) ? base.Text[..^1] : base.Text;
+            set => base.Text = $"{value}{_whiteSpace}";
+        }
 
         protected override void OnMouseDown(MouseEventArgs mevent)
         {
@@ -62,6 +71,13 @@ namespace EliteChroma.Controls
                 Brush brush = Enabled ? SystemBrushes.ControlText : SystemBrushes.ControlDark;
                 var arrows = new Point[] { new Point(arrowX, arrowY), new Point(arrowX + 7, arrowY), new Point(arrowX + 3, arrowY + 4) };
                 pevent.Graphics.FillPolygon(brush, arrows);
+
+                int dividerX = arrowX - 6;
+                int dividerY0 = 4;
+                int dividerY1 = ClientRectangle.Height - 5;
+
+                Pen pen = SystemPens.ControlLight;
+                pevent.Graphics.DrawLine(pen, new Point(dividerX, dividerY0), new Point(dividerX, dividerY1));
             }
         }
     }