+++ src/EliteChroma/Forms/Pages/PageGeneral.cs
@@ -1,0 +1,16 @@
+﻿namespace EliteChroma.Forms.Pages
+{
+    internal partial class PageGeneral : UserControl
+    {
+        public PageGeneral()
+        {
+            InitializeComponent();
+        }
+
+        public bool DetectGameInForeground
+        {
+            get => chDetectGameProcess.Checked;
+            set => chDetectGameProcess.Checked = value;
+        }
+    }
+}